helper initialization only happens when Rails is present
This will allow us to test the SwitchUser module without being dependent on
rails being loaded.
